bench: move 'encode' and 'decode' benches to 'integration'
