+++ Cargo.toml
@@ -22,11 +22,7 @@ criterion = { version = "0.5.1", features = ["html_reports"] }
 tempfile = "3.8.1"
 
 [[bench]]
-name = "encoding"
-harness = false
-
-[[bench]]
-name = "decoding"
+name = "integration"
 harness = false
 
 [[bench]]
